% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/smDropdown.R
\name{smDropdown}
\alias{smDropdown}
\title{smDropdown}
\usage{
smDropdown(parent, ..., el.list)
}
\arguments{
\item{parent}{A shiny object to attach the dropdown to. Usually a button or link.}

\item{el.list}{An optional \code{list} of elements to add to the dropdown}

\item{\dots}{\link{Elements} to add to the dropdown menu.}
}
\description{
A wrapper for attaching a dropdown menu to a shiny element.
}
\details{
Supplied with a \code{parent} and \link{Elements}, \code{smDropdown}
will attach a dropdown menu to \code{parent} that contains the provided
\link{Elements}.
}
\examples{
smDropdown(parent = actionButton("ddEx", "Dropdown"),
 smAction("act1", "An Action"),
 smAction("act2", "Another Action"),
 smDivider(),
 smCheckbox("cbEx", "cb1", "A Checkbox"),
 smCheckbox("cbEx", "cb2", "Another Checkbox")
)
}

